refactor: Phase 1.12 - Events 아키텍처 통합 및 레이어 정리
## 주요 변경사항
- 폴더 이름 변경: events/eventBus/ → events/bus/ (명칭 일관성)
- 파일 이름 변경: eventHandler.ts → discordEventHandler.ts (도메인 명시화)
- 인터페이스 구분: EventHandler<T>(EventBus) vs DiscordEventHandler(Discord.js)
- 파일 이동: fullActionId.ts → features/alarmSubscribe/constants/ (기능별 구조)
- 파일 제거: logHandler.ts (미사용, 정리)
- Import 경로 업데이트: 11개 파일

## 개선 효과
- ✅ Events 레이어 구조 명확화 (bus, handlers, listeners)
- ✅ 네이밍 충돌 제거 (EventHandler 구분)
- ✅ 기능별 구조 강화 (상수를 기능 폴더로 이동)
- ✅ Import 경로 단순화 (@/events/* 만 필요)
- ✅ 기술부채 감소 (불필요한 파일 정리)

Ref: .claude/todos/reviews/phase-1-12-review.md
