+++ functions/src/events/bus/EventBus.ts
@@ -10,7 +10,7 @@ import { eventLogger } from './utils/eventLogger';
  *
  * @example
  * ```typescript
- * import { eventBus, EventType } from '@/events/eventBus';
+ * import { eventBus, EventType } from '@/events';
  *
  * // 이벤트 발행
  * eventBus.emitEvent(EventType.RECRUIT_NEW, {
