+++ functions/src/events/bus/utils/eventLogger.ts
