+++ functions/src/events/bus/utils/registerEventHandlers.ts
@@ -27,7 +27,7 @@ let handlersRegistered = false;
  * @example
  * ```typescript
  * // app/index.ts에서 호출
- * import { registerAllEventHandlers } from '@/events/eventBus/utils/registerEventHandlers';
+ * import { registerAllEventHandlers } from '@/events';
  *
  * registerAllEventHandlers();
  * ```
