+++ functions/src/events/discordEventHandler.ts
@@ -1,6 +1,6 @@
 import type { ClientEvents } from "discord.js";
 
-export interface EventHandler {
+export interface DiscordEventHandler {
   event: keyof ClientEvents;
   once?: boolean;
   execute: (...args: any[]) => void | Promise<void>;
