+++ functions/src/features/alarmSubscribe/constants/fullActionId.ts
@@ -7,7 +7,7 @@ import {
   alertRegionEditActionId, 
   regionModeChangeActionId, 
   subscribeOptionActionId
-} from "@/features/alarmSubscribe/constants";
+} from "../constants";
 
 // TODO: 액션 ID 관리 용이성을 위해 리팩토링 고려 - prefix 별로 분리해놓고 합치는 방식
